Move microsoft-teams reducer and service to utils
