+++ utils/microsoft-teams/reducer.ts
