+++ utils/microsoft-teams/service/authorize.ts
